feat: ticker default command
