+++ src/commands/defi/ticker_slash/compare_slash.ts
