+++ src/commands/defi/token/index.ts
@@ -5,7 +5,7 @@ import add from "./add"
 import remove from "./remove"
 import list from "./list"
 import addcustom from "./addCustom"
-import compare from "./compare"
+import compare from "../ticker/compare"
 import setDefault from "./default"
 import info from "./info"
 import { thumbnails } from "utils/common"
